chore(ui): shared Pro-pill helper + replaceChildren in login-modal (#253, partial)
Addresses two of the three hygiene items in #253; the el() factory +
palette-constants migration are bigger and deferred to a follow-up.

## createProPill helper

The "Pro" pill shown on Pro-gated affordances was duplicated nearly
verbatim in two places:

- `src/ui/panel.ts` — mode-toggle badge.
- `src/ui/notebook-workspace.ts` — Insert-link-to-this-view badge.

~15 identical lines each. Extracted `createProPill(testid)` to
`src/ui/styles.ts` so the styling lives once and call sites just pick
a `data-testid`. Existing tests that query by `panel-mode-pro` /
`notebook-insert-link-pro` keep working. Unifies the two minor visual
drifts (8 vs 10px border-radius, 9 vs 10px font-size) onto one
size — imperceptible in practice, and one place to edit next time the
branding gets a tweak.

## replaceChildren in login-modal.ts

`body.innerHTML = ""` / `sent.innerHTML = ""` swapped for
`.replaceChildren()` — matches the pattern used in every other UI
module (panel.ts, drawer.ts, command-palette.ts). Forces the HTML
parser one less time per render cycle and closes a latent XSS vector
if user content ever flows through those nodes.

## Deferred

The broader `el()` factory and palette-constants migration land as a
follow-up — they touch 47 UI files and are opportunistic rather than
blocking. #253 stays open for that slice; re-titled once merged.

No behaviour change. 1221 SPA + 89 worker tests unchanged.

Full canonical gate green:
- typecheck (SPA + Worker)
- lint, format:check
- 1221 SPA tests, 89 Worker tests
- build
